api/requirements-streamlit.txt
# Streamlit Application Requirements

# Streamlit
streamlit==1.29.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Visualization
plotly==5.18.0

# API Calls
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Azure Application Insights (optionnel - pour monitoring)
opencensus-ext-azure==1.1.9
